TradeOS/requirements-local.txt
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.40
pywebview>=5.3
click>=8.1
requests>=2.31
yfinance>=0.2.54
pydantic>=2.0,<3.0
pydantic-settings>=2.0
pyyaml>=6.0
structlog>=24.0
python-dateutil>=2.8
tenacity>=8.0
orjson>=3.9
pandas>=2.0
numpy>=1.24
